chore: move ConfigItem to src folder
