+++ src/ConfigItem.ts
